requirements.txt
llama-index
llama-parse
openai
gradio
python-dotenv
nest_asyncio
app
